fix: fix test

+++ test/server/comm/guicommchannel/testerrorjobs.cpp
@@ -221,7 +221,7 @@ void TestGuiCommChannel::testErrorDeleteJob() {
     auto processFct = [](std::shared_ptr<AbstractGuiJob> job) {
         auto deleteJob = std::dynamic_pointer_cast<ErrorDeleteJob>(job);
         CPPUNIT_ASSERT(deleteJob);
-        CPPUNIT_ASSERT_EQUAL(42, deleteJob->_errorDbId);
+        CPPUNIT_ASSERT_EQUAL(int64_t(42), deleteJob->_errorDbId);
     };
 
 #if defined(KD_WINDOWS) || defined(KD_LINUX)
